Attempt to support newer SDK version?
I did this work a couple years ago and I don't recall exactly what I was
doing. It looks like I tried to update the source code such that it is
more compatible with the SDK version that was current at the time, which
I added in the previous commit. I'm not sure if I managed to get things
working with this newer SDK or not.